--- src/main/java/com/practice/course_registration/Role.java
@@ -1,5 +1,0 @@
-package com.practice.course_registration;
-
-public enum Role {
-    STUDENT, PROFESSOR, ADMIN, AUTH_STUDENT;
-}
