--- src/main/java/com/practice/course_registration/apiPayload/code/BaseCode.java
@@ -1,10 +1,0 @@
-package com.practice.course_registration.apiPayload.code;
-
-
-
-public interface BaseCode {
-
-    ResponseDTO getReason();
-
-    ResponseDTO getReasonHttpStatus();
-}
