--- src/main/java/com/practice/course_registration/apiPayload/code/BaseErrorCode.java
@@ -1,10 +1,0 @@
-package com.practice.course_registration.apiPayload.code;
-
-
-public interface BaseErrorCode {
-
-    ResponseDTO getReason();
-
-    ResponseDTO getReasonHttpStatus();
-
-}
