--- src/main/java/com/practice/course_registration/controller/LoginController.java
@@ -1,14 +1,0 @@
-package com.practice.course_registration.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-    @GetMapping("/login")
-    public String loginPage(){
-        return "login";
-    }
-
-
-}
