--- src/main/java/com/practice/course_registration/controller/MainController.java
@@ -1,13 +1,0 @@
-package com.practice.course_registration.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-    @GetMapping("/")
-    public String main(){
-        return "main";
-    }
-}
-
