+++ src/main/java/com/practice/course_registration/domain/member/controller/LoginController.java
@@ -1,0 +1,14 @@
+package com.practice.course_registration.domain.member.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginController {
+    @GetMapping("/login")
+    public String loginPage(){
+        return "login";
+    }
+
+
+}
