+++ src/main/java/com/practice/course_registration/domain/member/controller/MainController.java
@@ -1,0 +1,13 @@
+package com.practice.course_registration.domain.member.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+    @GetMapping("/")
+    public String main(){
+        return "main";
+    }
+}
+
