+++ src/main/java/com/practice/course_registration/global/apiPayload/ApiCommonResponse.java
@@ -1,0 +1,35 @@
+package com.practice.course_registration.global.apiPayload;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import com.practice.course_registration.global.apiPayload.code.status.SuccessStatus;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+@JsonPropertyOrder({"isSuccess", "code", "message", "result"})
+public class ApiCommonResponse<T> {
+
+    @JsonProperty("isSuccess")
+    private final Boolean isSuccess;
+
+    private final String code;
+
+    private final String message;
+
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private T result;
+
+
+    // 성공시 응답
+    public static <T> ApiCommonResponse<T> onSuccess(T result) {
+        return new ApiCommonResponse<>(true, SuccessStatus._OK.getCode() , SuccessStatus._OK.getMessage(), result);
+    }
+
+    public static <T> ApiCommonResponse<T> onFailure(String code, String message, T data) {
+        return new ApiCommonResponse<>(false, code, message, data);
+    }
+
+}
