+++ src/main/java/com/practice/course_registration/global/apiPayload/code/BaseCode.java
@@ -1,0 +1,10 @@
+package com.practice.course_registration.global.apiPayload.code;
+
+
+
+public interface BaseCode {
+
+    ResponseDTO getReason();
+
+    ResponseDTO getReasonHttpStatus();
+}
