+++ src/main/java/com/practice/course_registration/global/apiPayload/code/BaseErrorCode.java
@@ -1,0 +1,10 @@
+package com.practice.course_registration.global.apiPayload.code;
+
+
+public interface BaseErrorCode {
+
+    ResponseDTO getReason();
+
+    ResponseDTO getReasonHttpStatus();
+
+}
