+++ src/main/java/com/practice/course_registration/global/apiPayload/code/ResponseDTO.java
@@ -1,0 +1,20 @@
+package com.practice.course_registration.global.apiPayload.code;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Builder
+public class ResponseDTO {
+
+    private HttpStatus httpStatus;
+
+    private final boolean isSuccess;
+    private final String code;
+    private final String message;
+
+    public boolean getIsSuccess() {
+        return isSuccess;
+    }
+}
