+++ src/main/java/com/practice/course_registration/global/apiPayload/exception/GeneralException.java
@@ -1,0 +1,21 @@
+package com.practice.course_registration.global.apiPayload.exception;
+
+import com.practice.course_registration.global.apiPayload.code.BaseErrorCode;
+import com.practice.course_registration.global.apiPayload.code.ResponseDTO;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class GeneralException extends RuntimeException {
+
+    private BaseErrorCode code;
+
+    public ResponseDTO getErrorReason() {
+        return this.code.getReason();
+    }
+
+    public ResponseDTO getErrorReasonHttpStatus(){
+        return this.code.getReasonHttpStatus();
+    }
+}
