+++ src/main/java/com/practice/course_registration/global/enums/Role.java
@@ -1,0 +1,5 @@
+package com.practice.course_registration.global.enums;
+
+public enum Role {
+    STUDENT, PROFESSOR, ADMIN, AUTH_STUDENT;
+}
