--- src/main/java/com/practice/course_registration/repository/MemberRepository.java
@@ -1,11 +1,0 @@
-package com.practice.course_registration.repository;
-
-import com.practice.course_registration.entity.MemberEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
-    boolean existsByMemberId(String memberId);
-    boolean existsByMemberNumber(String memberNumber);
-
-    MemberEntity findByMemberId(String memberId);
-}
